# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "registerUser"
                ],
                "summary": "Регистрация пользователя",
                "description": "Регистрация пользователя",
                "operationId": "registerUser",
                "requestBody": {
                    "description": "Параметры",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "bahrom"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "bahrom.ermatov@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Observer1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Пользователь успешно зарегистрирован"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при регистрации водителя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Возникла ошибка при регистрации пользоватея"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "loginUser"
                ],
                "summary": "Авторизация пользователя",
                "description": "Авторизация пользователя",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "Параметры",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "bahrom.ermatov@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Observer1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Пользователь успешно авторизовался"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при регистрации водителя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Возникла ошибка при авторизации пользоватея"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/driver/select/{amount}/{latitude}/{longitude}": {
            "get": {
                "tags": [
                    "selectDriver"
                ],
                "summary": "Выбор водителя",
                "description": "Выбор водителя",
                "operationId": "selectDriver",
                "parameters": [
                    {
                        "name": "amount",
                        "in": "path",
                        "description": "Сумма заказа",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "100"
                    },
                    {
                        "name": "latitude",
                        "in": "path",
                        "description": "Latitude",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "40.304213"
                    },
                    {
                        "name": "longitude",
                        "in": "path",
                        "description": "Longitude",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "69.632049"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при выборе водителя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Возникла ошибка при выборе водителя"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/driver/charge-commission": {
            "post": {
                "tags": [
                    "chargeCommission"
                ],
                "summary": "Биллинг - Начисление коммиссии",
                "description": "Биллинг - Начисление коммиссии",
                "operationId": "chargeCommission",
                "requestBody": {
                    "description": "Параметры",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "driver_id",
                                    "amount",
                                    "commission"
                                ],
                                "properties": {
                                    "driver_id": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": "1"
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": "100"
                                    },
                                    "commission": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Коммиссия успешно начислена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при начислении коммиссии",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Возникла ошибка при начислении коммиссии"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/driver/register": {
            "post": {
                "tags": [
                    "registerDriver"
                ],
                "summary": "Регистрация водителя",
                "description": "Регистрация водителя",
                "operationId": "registerDriver",
                "requestBody": {
                    "description": "Параметры",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "login",
                                    "balance",
                                    "latitude",
                                    "longitude",
                                    "rating"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Bahrom"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Ermatov"
                                    },
                                    "login": {
                                        "type": "string",
                                        "example": "bahrom"
                                    },
                                    "balance": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "0"
                                    },
                                    "latitude": {
                                        "type": "string",
                                        "example": "40.289541"
                                    },
                                    "longitude": {
                                        "type": "string",
                                        "example": "69.632049"
                                    },
                                    "rating": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "0.5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Водитель успешно зарегистрирован"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при регистрации водителя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Возникла ошибка при регистрации водителя"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    ]
}